Factor out decodeBlock method from decodeArray
When using OptPfor and NewPfor in inverted index compression, it is
often useful to have more control over the space overhead caused by the
encoding and decoding process as the size of the encoded integers are
already stored elsewhere or, integers are compressed in fixed blocks of
a certain size. Unfortunatelly, when using the encodeArray and
decodeArray interface of the library (at least in my experience with
OptPfor) the size of the encoded data is always stored explicitly.
There is already a way to use the encodeBlock() function explicitly to
avoid this overhead during compression time, however the corresponding
decodeBlock method does not exist.

This commit factors out the "decodeBlock" functionality in the
decodeArray function of NewPfor into a new function which can then be
used in conjunction with encodeBlock(). All unit tests still pass and
the decoding code essentially stays the same.